+++ lambdas/data-migration1/src/errors.ts
@@ -1,0 +1,3 @@
+import { createErrorType } from '@cumulus/errors';
+
+export const RecordAlreadyMigrated = createErrorType('RecordAlreadyMigrated');
