+++ lambdas/data-migration1/src/types.ts
@@ -1,0 +1,6 @@
+export interface MigrationSummary {
+  dynamoRecords: number
+  success: number
+  skipped: number
+  failed: number
+}
